# openapi.yaml
openapi: 3.0.0
paths:
  /auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/AuthResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Register new user
      tags: &ref_0
        - Authentication
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/AuthResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Login user
      tags: *ref_0
...
  /auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: object
                    nullable: true
                    example: null
                    description: No data returned
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      security:
        - bearer: []
      summary: Logout user (client-side token removal)
      tags: *ref_0
  /auth/send-verification:
    post:
      operationId: AuthController_sendVerification
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendVerificationDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/VerificationCodeResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Send verification code to email
      tags: *ref_0
  /auth/confirm-verification:
    post:
      operationId: AuthController_confirmVerification
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmVerificationDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: object
                    nullable: true
                    example: null
                    description: No data returned
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Confirm verification code
      tags: *ref_0
  /auth/reset-password:
    post:
      operationId: AuthController_resetPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: object
                    nullable: true
                    example: null
                    description: No data returned
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      summary: Reset password using verification code
      tags: *ref_0
  /auth/change-password:
    post:
      operationId: AuthController_changePassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: object
                    nullable: true
                    example: null
                    description: No data returned
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      security:
        - bearer: []
      summary: Change password (authenticated user)
      tags: *ref_0
  /auth/me:
    post:
      operationId: AuthController_getCurrentUser
      parameters: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/UserInfoDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      security:
        - bearer: []
      summary: Get current authenticated user
      tags: *ref_0
...
info:
  title: API Docs
  description: Custom Swagger UI in NestJS
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UserInfoDto:
      type: object
      properties:
        id:
          type: string
          example: user_123
        email:
          type: string
          example: user@example.com
        username:
          type: string
          example: john_doe
        role:
          type: string
          example: FAN
      required:
        - id
        - email
        - username
        - role
    AuthResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/UserInfoDto'
      required:
        - accessToken
        - refreshToken
        - user
...
    VerificationCodeResponseDto:
      type: object
      properties:
        expiresAt:
          type: number
          description: Expiration timestamp
          example: 1740412406898
      required:
        - expiresAt
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        username:
          type: string
          example: john_doe
        password:
          type: string
          example: SecurePass123!
        role:
          type: string
          example: FAN
          enum:
            - FAN
            - IDOL
      required:
        - email
        - username
        - password
    LoginDto:
      type: object
      properties:
        credential:
          type: string
          example: user@example.com or john_doe
        password:
          type: string
          example: SecurePass123!
      required:
        - credential
        - password
    SendVerificationDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        verificationType:
          type: string
          enum:
            - REGISTER_ACCOUNT
            - RESET_PASSWORD
            - CHANGE_PASSWORD
            - FIND_EMAIL
          example: REGISTER_ACCOUNT
      required:
        - email
        - verificationType
    ConfirmVerificationDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        code:
          type: string
          example: '123456'
        verificationType:
          type: string
          enum:
            - REGISTER_ACCOUNT
            - RESET_PASSWORD
            - CHANGE_PASSWORD
            - FIND_EMAIL
          example: REGISTER_ACCOUNT
      required:
        - email
        - code
        - verificationType
    ResetPasswordDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: NewSecurePass123!
      required:
        - email
        - password
    ChangePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          example: OldSecurePass123!
        password:
          type: string
          example: NewSecurePass123!
      required:
        - oldPassword
        - password
...